Queue up closed pull requests for processing
Without this patch the pull request controller discards closed pull
requests.  This is a problem because we need to pull the most recent
comment from the pull request and update the Trello card when the pull
request is closed.  This patch addresses the problem by creating a pull
request job and putting it into the delayed job queue.

The delayed job will need to decide what to do with the closed pull
request.