Add a log message when a card already exists
Without this patch the app does not let the user know when a job does
not take any action because a found card already exists.  The problem
happens in all jobs that implement the default behavior of the
BaseTrelloJob.

This patch addresses the problem by logging a message with the card
short id and URL.